requirements.txt
fastapi
uvicorn
aiosqlite
sse-starlette
google-genai
openai
ddgs
httpx
trafilatura
beautifulsoup4
python-dotenv
pypdf
python-docx
pandas
faiss-cpu
sentence-transformers
anthropic
pyyaml
